Rename Find_Indices to Find_Indices.py
Intuition
I initially considered checking if the difference between the target and each number is present in the array, rather than using nested loops.

Approach
I formed a loop that goes through the whole array calculating the difference (diff) and asked the system if that number besides the current iteration of the number is in the list. If it is, then record the both the numbers.